requirements.txt
python-telegram-bot
requests
openai
pydub
textract
